fix: tipos

+++ include/Pipa.hpp
@@ -5,7 +5,7 @@
 
 class Pipa : public Coletavel {
     public:
-        Pipa(float pos_x, float pos_y);
+        Pipa(double pos_x, double pos_y);
         ~Pipa();
         void aplicarEfeito(Jogador& jogador) override;
 };
