+++ src/projetil.cpp
@@ -1,6 +1,6 @@
 #include "Projetil.hpp"
 
-Projetil::Projetil(float pos_x, float pos_y, float vel_x, float vel_y, int dano){
+Projetil::Projetil(double pos_x, double pos_y, double vel_x, double vel_y, int dano) {
     this->pos_x = pos_x;
     this->pos_y = pos_y;
     this->vel_x = vel_x;
@@ -9,25 +9,24 @@ Projetil::Projetil(float pos_x, float pos_y, float vel_x, float vel_y, int dano)
     this->ativo = true;
 }
 
-Projetil::~Projetil(){
-}
+Projetil::~Projetil() {}
 
-int Projetil::getDano() const{
+int Projetil::getDano() const {
     return(dano);
 }
 
-float Projetil::getPosX() const{
+double Projetil::getPosX() const {
     return(pos_x);
 }
 
-float Projetil::getPosY() const{
+double Projetil::getPosY() const {
     return(pos_y);
 }
 
-bool Projetil::isAtivo() const{
+bool Projetil::isAtivo() const {
     return(ativo);
 }
 
-void Projetil::desativar(){
+void Projetil::desativar() {
     this->ativo = false;
 }
